move shard interfaces to shard.go
